Added first CPU instruction: ADC.
Changes include:
- Created an opcode/instruction table filled with NOPs for now.
- Added the ADC instructions to the instruction table.
- Implemented addressing modes (immediate/implicit/absolute/etc.) that
  are used by the ADC instruction.
- Added tests to verify the behavior and cycle accuracy of the ADC
  instruction.
- Added compiler warning options for MSVC and gcc to ensure the highest
  warning level (which includes erroring when an warning is encountered
  with gcc).
- Made the default CMake build mode 'Release' if it is not specified on
  the command line.

+++ include/nesturbia/types.hpp
@@ -16,7 +16,7 @@ template <unsigned Precision> struct uint {
   operator type_t() const { return data; }
 
   [[nodiscard]] bool bit(int bit) const {
-    if (bit < 0 || bit >= (sizeof(type_t) * 8)) {
+    if (bit < 0 || bit >= static_cast<int>((sizeof(type_t) * 8))) {
       return false;
     }
